Email Auto-Categorizer_Buvaneswari/email_auto_categorizer/requirements_live.txt
streamlit
pandas
scikit-learn
numpy
PyPDF2
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
schedule
fastapi
uvicorn
requests
plotly
